Moved the Create activity handler to a separate file
refs https://linear.app/ghost/issue/AP-718

We still have quite a few deps that we have to import, but those will go away
as we migrate towards the feeds table. This is the first step in breaking up
the large dispatchers.ts file into more managable chunks.
